Crate rustc_target

Some stuff used by rustc that doesn’t have many dependencies

Originally extracted from rustc::back, which was nominally the compiler ‘backend’, though LLVM is rustc’s backend, so rustc_target is really just odds-and-ends relating to code gen and linking. This crate mostly exists to make rustc smaller, so we might put more ‘stuff’ here in the future. It does not have a dependency on LLVM.

Modules

abi

asm

callconv

json

spec

Flexible target specification.

target_features

Declares Rust’s target feature names for each target. Note that these are similar to but not always identical to LLVM’s feature names, and Rust adds some features that do not correspond to LLVM features at all.

Constants

RUST_LIB_DIR 🔒

The name of rustc’s own place to organize libraries.

Traits

HashStableContext

Requirements for a StableHashingContext to be used in this crate. This is a hack to allow using the HashStable_Generic derive macro instead of implementing everything in rustc_middle.

Functions

find_relative_libdir 🔒

The name of the directory rustc expects libraries to be located.

relative_target_rustlib_path

Returns a rustlib path for this particular target, relative to the provided sysroot.