Module rustc_target::abi::call

Structs

ArgAttribute

ArgAttributes
A compact representation of LLVM attributes (at least those relevant for this module) that can be manipulated without interacting with LLVM's Attribute machinery.

ArgType
Information about how to pass an argument to, or return a value from, a function, under some ABI.

CastTarget

FnType
Metadata describing how the arguments to a native function should be passed in order to respect the native ABI.

Reg

Uniform
An argument passed entirely registers with the same kind (e.g., HFA / HVA on PPC64 and AArch64).

Enums

Conv

HomogeneousAggregate
Returns value from the homogeneous_aggregate test function.

IgnoreMode

PassMode

RegKind