Struct rustc_target::abi::call::ArgAttributes

pub struct ArgAttributes {
    pub regular: ArgAttribute,
    pub arg_ext: ArgExtension,
    pub pointee_size: Size,
    pub pointee_align: Option<Align>,
}

A compact representation of LLVM attributes (at least those relevant for this module) that can be manipulated without interacting with LLVM’s Attribute machinery.

Fields
regular: ArgAttributearg_ext: ArgExtensionpointee_size: Size

The minimum size of the pointee, guaranteed to be valid for the duration of the whole call (corresponding to LLVM’s dereferenceable and dereferenceable_or_null attributes).

pointee_align: Option<Align>
Implementations

impl ArgAttributes

pub fn new() -> Self

pub fn ext(&mut self, ext: ArgExtension) -> &mut Self

pub fn set(&mut self, attr: ArgAttribute) -> &mut Self

pub fn contains(&self, attr: ArgAttribute) -> bool