Struct rustc_target::abi::call::FnAbi

pub struct FnAbi<'a, Ty> {
    pub args: Box<[ArgAbi<'a, Ty>]>,
    pub ret: ArgAbi<'a, Ty>,
    pub c_variadic: bool,
    pub fixed_count: u32,
    pub conv: Conv,
    pub can_unwind: bool,
}

Metadata describing how the arguments to a native function should be passed in order to respect the native ABI.

I will do my best to describe this structure, but these comments are reverse-engineered and may be inaccurate. -NDM

Fields
args: Box<[ArgAbi<'a, Ty>]>

The LLVM types of each argument.

ret: ArgAbi<'a, Ty>

LLVM return type.

c_variadic: boolfixed_count: u32

The count of non-variadic arguments.

Should only be different from args.len() when c_variadic is true. This can be used to know whether an argument is variadic or not.

conv: Convcan_unwind: bool
Implementations

impl<'a, Ty> FnAbi<'a, Ty>

pub fn adjust_for_foreign_abi<C>(
&mut self,
cx: &C,
abi: Abi
) -> Result<(), AdjustForForeignAbiError>where
Ty: TyAbiInterface<'a, C> + Copy,
C: HasDataLayout + HasTargetSpec,