Enum rustc_target::abi::FieldsShape

pub enum FieldsShape {
    Primitive,
    Union(NonZeroUsize),
    Array {
        stride: Size,
        count: u64,
    },
    Arbitrary {
        offsets: Vec<Size>,
        memory_index: Vec<u32>,
    },
}

Describes how the fields of a type are located in memory.

Variants

Primitive

Scalar primitives and !, which never have fields.

Union(NonZeroUsize)

All fields start at no offset. The usize is the field count.

Array

Fields

stride: Size

count: u64

Array/vector-like placement, with all fields of identical types.

Arbitrary

Fields

offsets: Vec<Size>

Offsets for the first byte of each field, ordered to match the source definition order. This vector does not go in increasing order.

memory_index: Vec<u32>

Maps source order field indices to memory order indices, depending on how the fields were reordered (if at all). This is a permutation, with both the source order and the memory order using the same (0..n) index ranges.

Note that during computation of memory_index, sometimes it is easier to operate on the inverse mapping (that is, from memory order to source order), and that is usually named inverse_memory_index.

Struct-like placement, with precomputed offsets.

Fields are guaranteed to not overlap, but note that gaps before, between and after all the fields are NOT always padding, and as such their contents may not be discarded. For example, enum variants leave a gap at the start, where the discriminant field in the enum layout goes.

Implementations

impl FieldsShape

pub fn count(&self) -> usize

pub fn offset(&self, i: usize) -> Size

pub fn memory_index(&self, i: usize) -> usize

pub fn index_by_increasing_offset<'a>(
&'a self
) -> impl Iterator<Item = usize> + 'a

Gets source indices of the fields by increasing offsets.