Struct rustc_target::abi::TyAndLayout

pub struct TyAndLayout<'a, Ty> {
    pub ty: Ty,
    pub layout: Layout<'a>,
}

The layout of a type, alongside the type itself. Provides various type traversal APIs (e.g., recursing into fields).

Note that the layout is NOT guaranteed to always be identical to that obtained from layout_of(ty), as we need to produce layouts for which Rust types do not exist, such as enum variants or synthetic fields of enums (i.e., discriminants) and fat pointers.

Fields
ty: Tylayout: Layout<'a>
Implementations

impl<'a, Ty> TyAndLayout<'a, Ty>

fn is_aggregate(&self) -> bool

pub fn homogeneous_aggregate<C>(
&self,
cx: &C
) -> Result<HomogeneousAggregate, Heterogeneous>where
Ty: TyAbiInterface<'a, C> + Copy,

Returns Homogeneous if this layout is an aggregate containing fields of only a single type (e.g., (u32, u32)). Such aggregates are often special-cased in ABIs.

Note: We generally ignore fields of zero-sized type when computing this value (see #56877).

This is public so that it can be used in unit tests, but should generally only be relevant to the ABI details of specific targets.

impl<'a, Ty> TyAndLayout<'a, Ty>

pub fn for_variant<C>(self, cx: &C, variant_index: VariantIdx) -> Selfwhere
Ty: TyAbiInterface<'a, C>,

pub fn field<C>(self, cx: &C, i: usize) -> Selfwhere
Ty: TyAbiInterface<'a, C>,

pub fn pointee_info_at<C>(self, cx: &C, offset: Size) -> Option<PointeeInfo>where
Ty: TyAbiInterface<'a, C>,

pub fn is_single_fp_element<C>(self, cx: &C) -> boolwhere
Ty: TyAbiInterface<'a, C>,
C: HasDataLayout,

pub fn is_adt<C>(self) -> boolwhere
Ty: TyAbiInterface<'a, C>,

pub fn is_never<C>(self) -> boolwhere
Ty: TyAbiInterface<'a, C>,

pub fn is_tuple<C>(self) -> boolwhere
Ty: TyAbiInterface<'a, C>,

pub fn is_unit<C>(self) -> boolwhere
Ty: TyAbiInterface<'a, C>,

impl<'a, Ty> TyAndLayout<'a, Ty>

pub fn is_unsized(&self) -> bool

Returns true if the layout corresponds to an unsized type.

pub fn is_zst(&self) -> bool

Returns true if the type is a ZST and not unsized.

pub fn might_permit_raw_init<C>(self, cx: &C, init_kind: InitKind) -> boolwhere
Self: Copy,
Ty: TyAbiInterface<'a, C>,
C: HasDataLayout,

Determines if this type permits “raw” initialization by just transmuting some memory into an instance of T.

init_kind indicates if the memory is zero-initialized or left uninitialized.

This code is intentionally conservative, and will not detect

zero init of an enum whose 0 variant does not allow zero initialization

making uninitialized types who have a full valid range (ints, floats, raw pointers)

Any form of invalid value being made inside an array (unless the value is uninhabited)

A strict form of these checks that uses const evaluation exists in rustc_const_eval::might_permit_raw_init, and a tracking issue for making these checks stricter is https://github.com/rust-lang/rust/issues/66151.

FIXME: Once all the conservatism is removed from here, and the checks are ran by default, we can use the const evaluation checks always instead.