Struct rustc_target::abi::TyLayout

pub struct TyLayout<'a, Ty> {
    pub ty: Ty,
    pub details: &'a LayoutDetails,
}

The details of the layout of a type, alongside the type itself. Provides various type traversal APIs (e.g., recursing into fields).

Note that the details are NOT guaranteed to always be identical to those obtained from layout_of(ty), as we need to produce layouts for which Rust types do not exist, such as enum variants or synthetic fields of enums (i.e., discriminants) and fat pointers.

Fields
ty: Tydetails: &'a LayoutDetails
Methods

impl<'a, Ty> TyLayout<'a, Ty>

pub fn homogeneous_aggregate<C>(&self, cx: &C) -> HomogeneousAggregate where
Ty: TyLayoutMethods<'a, C> + Copy,
C: LayoutOf<Ty = Ty, TyLayout = Self>,

Returns true if this layout is an aggregate containing fields of only a single type (e.g., (u32, u32)). Such aggregates are often special-cased in ABIs.

Note: We generally ignore fields of zero-sized type when computing this value (see #56877).

This is public so that it can be used in unit tests, but should generally only be relevant to the ABI details of specific targets.

impl<'a, Ty> TyLayout<'a, Ty>

pub fn for_variant<C>(self, cx: &C, variant_index: VariantIdx) -> Self where
Ty: TyLayoutMethods<'a, C>,
C: LayoutOf<Ty = Ty>,

pub fn field<C>(self, cx: &C, i: usize) -> C::TyLayout where
Ty: TyLayoutMethods<'a, C>,
C: LayoutOf<Ty = Ty>,

impl<'a, Ty> TyLayout<'a, Ty>

pub fn is_unsized(&self) -> bool

Returns true if the layout corresponds to an unsized type.

pub fn is_zst(&self) -> bool

Returns true if the type is a ZST and not unsized.