Function rustc_target::asm::allocatable_registers

pub fn allocatable_registers(
    arch: InlineAsmArch,
    reloc_model: RelocModel,
    target_features: &FxHashSet<Symbol>,
    target: &Target
) -> FxHashMap<InlineAsmRegClass, FxHashSet<InlineAsmReg>>

Returns the full set of allocatable registers for a given architecture.

The registers are structured as a map containing the set of allocatable registers in each register class. A particular register may be allocatable from multiple register classes, in which case it will appear multiple times in the map.