rustc_target::spec::base

Module apple

Structs

OSVersion

Deployment target or SDK version.

Enums

Arch 🔒

TargetEnv 🔒

Functions

base 🔒

Get the base target options, unversioned LLVM target and target_arch from the three things that uniquely identify Rust’s Apple targets: The OS, the architecture, and the ABI.

deployment_target_env_var

Name of the environment variable used to fetch the deployment target on the given OS.

link_env_remove 🔒

unversioned_llvm_target 🔒

Generate part of the LLVM target triple.