Enum rustc_target::spec::StackProtector

pub enum StackProtector {
    None,
    Basic,
    Strong,
    All,
}

Controls use of stack canaries.

Variants

None

Disable stack canary generation.

Basic

On LLVM, mark all generated LLVM functions with the ssp attribute (see llvm/docs/LangRef.rst). This triggers stack canary generation in functions which contain an array of a byte-sized type with more than eight elements.

Strong

On LLVM, mark all generated LLVM functions with the sspstrong attribute (see llvm/docs/LangRef.rst). This triggers stack canary generation in functions which either contain an array, or which take the address of a local variable.

All

Generate stack canaries in all functions.

Implementations

impl StackProtector

fn as_str(&self) -> &'static str