Struct rustc_target::spec::TargetOptions

pub struct TargetOptions {
    pub is_builtin: bool,
    pub linker: Option<String>,
    pub lld_flavor: LldFlavor,
    pub pre_link_args: LinkArgs,
    pub pre_link_args_crt: LinkArgs,
    pub pre_link_objects_exe: Vec<String>,
    pub pre_link_objects_exe_crt: Vec<String>,
    pub pre_link_objects_dll: Vec<String>,
    pub late_link_args: LinkArgs,
    pub post_link_objects: Vec<String>,
    pub post_link_objects_crt: Vec<String>,
    pub post_link_args: LinkArgs,
    pub link_env: Vec<(String, String)>,
    pub asm_args: Vec<String>,
    pub cpu: String,
    pub features: String,
    pub dynamic_linking: bool,
    pub only_cdylib: bool,
    pub executables: bool,
    pub relocation_model: String,
    pub code_model: Option<String>,
    pub tls_model: String,
    pub disable_redzone: bool,
    pub eliminate_frame_pointer: bool,
    pub function_sections: bool,
    pub dll_prefix: String,
    pub dll_suffix: String,
    pub exe_suffix: String,
    pub staticlib_prefix: String,
    pub staticlib_suffix: String,
    pub target_family: Option<String>,
    pub abi_return_struct_as_int: bool,
    pub is_like_osx: bool,
    pub is_like_solaris: bool,
    pub is_like_windows: bool,
    pub is_like_msvc: bool,
    pub is_like_android: bool,
    pub is_like_emscripten: bool,
    pub is_like_fuchsia: bool,
    pub linker_is_gnu: bool,
    pub allows_weak_linkage: bool,
    pub has_rpath: bool,
    pub no_default_libraries: bool,
    pub position_independent_executables: bool,
    pub needs_plt: bool,
    pub relro_level: RelroLevel,
    pub archive_format: String,
    pub allow_asm: bool,
    pub custom_unwind_resume: bool,
    pub has_elf_tls: bool,
    pub obj_is_bitcode: bool,
    pub no_integrated_as: bool,
    pub min_atomic_width: Option<u64>,
    pub max_atomic_width: Option<u64>,
    pub atomic_cas: bool,
    pub panic_strategy: PanicStrategy,
    pub abi_blacklist: Vec<Abi>,
    pub crt_static_allows_dylibs: bool,
    pub crt_static_default: bool,
    pub crt_static_respected: bool,
    pub stack_probes: bool,
    pub min_global_align: Option<u64>,
    pub default_codegen_units: Option<u64>,
    pub trap_unreachable: bool,
    pub requires_lto: bool,
    pub singlethread: bool,
    pub no_builtins: bool,
    pub i128_lowering: bool,
    pub codegen_backend: String,
    pub default_hidden_visibility: bool,
    pub embed_bitcode: bool,
    pub emit_debug_gdb_scripts: bool,
    pub requires_uwtable: bool,
    pub simd_types_indirect: bool,
    pub override_export_symbols: Option<Vec<String>>,
    pub merge_functions: MergeFunctions,
}

Optional aspects of a target specification.

This has an implementation of Default, see each field for what the default is. In general, these try to take "minimal defaults" that don't assume anything about the runtime they run in.

Fields
is_builtin: bool

Whether the target is built-in or loaded from a custom target specification.

linker: Option<String>

Linker to invoke

lld_flavor: LldFlavor

LLD flavor

pre_link_args: LinkArgs

Linker arguments that are passed before any user-defined libraries.

pre_link_args_crt: LinkArgspre_link_objects_exe: Vec<String>

Objects to link before all others, always found within the sysroot folder.

pre_link_objects_exe_crt: Vec<String>pre_link_objects_dll: Vec<String>late_link_args: LinkArgs

Linker arguments that are unconditionally passed after any user-defined but before post_link_objects. Standard platform libraries that should be always be linked to, usually go here.

post_link_objects: Vec<String>

Objects to link after all others, always found within the sysroot folder.

post_link_objects_crt: Vec<String>post_link_args: LinkArgs

Linker arguments that are unconditionally passed after any user-defined libraries.

link_env: Vec<(String, String)>

Environment variables to be set before invoking the linker.

asm_args: Vec<String>

Extra arguments to pass to the external assembler (when used)

cpu: String

Default CPU to pass to LLVM. Corresponds to llc -mcpu=$cpu. Defaults to "generic".

features: String

Default target features to pass to LLVM. These features will always be passed, and cannot be disabled even via -C. Corresponds to llc -mattr=$features.

dynamic_linking: bool

Whether dynamic linking is available on this target. Defaults to false.

only_cdylib: bool

If dynamic linking is available, whether only cdylibs are supported.

executables: bool

Whether executables are available on this target. iOS, for example, only allows static libraries. Defaults to false.

relocation_model: String

Relocation model to use in object file. Corresponds to llc -relocation-model=$relocation_model. Defaults to "pic".

code_model: Option<String>

Code model to use. Corresponds to llc -code-model=$code_model.

tls_model: String

TLS model to use. Options are "global-dynamic" (default), "local-dynamic", "initial-exec" and "local-exec". This is similar to the -ftls-model option in GCC/Clang.

disable_redzone: bool

Do not emit code that uses the "red zone", if the ABI has one. Defaults to false.

eliminate_frame_pointer: bool

Eliminate frame pointers from stack frames if possible. Defaults to true.

function_sections: bool

Emit each function in its own section. Defaults to true.

dll_prefix: String

String to prepend to the name of every dynamic library. Defaults to "lib".

dll_suffix: String

String to append to the name of every dynamic library. Defaults to ".so".

exe_suffix: String

String to append to the name of every executable.

staticlib_prefix: String

String to prepend to the name of every static library. Defaults to "lib".

staticlib_suffix: String

String to append to the name of every static library. Defaults to ".a".

target_family: Option<String>

OS family to use for conditional compilation. Valid options: "unix", "windows".

abi_return_struct_as_int: bool

Whether the target toolchain's ABI supports returning small structs as an integer.

is_like_osx: bool

Whether the target toolchain is like macOS's. Only useful for compiling against iOS/macOS, in particular running dsymutil and some other stuff like -dead_strip. Defaults to false.

is_like_solaris: bool

Whether the target toolchain is like Solaris's. Only useful for compiling against Illumos/Solaris, as they have a different set of linker flags. Defaults to false.

is_like_windows: bool

Whether the target toolchain is like Windows'. Only useful for compiling against Windows, only really used for figuring out how to find libraries, since Windows uses its own library naming convention. Defaults to false.

is_like_msvc: boolis_like_android: bool

Whether the target toolchain is like Android's. Only useful for compiling against Android. Defaults to false.

is_like_emscripten: bool

Whether the target toolchain is like Emscripten's. Only useful for compiling with Emscripten toolchain. Defaults to false.

is_like_fuchsia: bool

Whether the target toolchain is like Fuchsia's.

linker_is_gnu: bool

Whether the linker support GNU-like arguments such as -O. Defaults to false.

allows_weak_linkage: bool

The MinGW toolchain has a known issue that prevents it from correctly handling COFF object files with more than 215 sections. Since each weak symbol needs its own COMDAT section, weak linkage implies a large number sections that easily exceeds the given limit for larger codebases. Consequently we want a way to disallow weak linkage on some platforms.

has_rpath: bool

Whether the linker support rpaths or not. Defaults to false.

no_default_libraries: bool

Whether to disable linking to the default libraries, typically corresponds to -nodefaultlibs. Defaults to true.

position_independent_executables: bool

Dynamically linked executables can be compiled as position independent if the default relocation model of position independent code is not changed. This is a requirement to take advantage of ASLR, as otherwise the functions in the executable are not randomized and can be used during an exploit of a vulnerability in any code.

needs_plt: bool

Determines if the target always requires using the PLT for indirect library calls or not. This controls the default value of the -Z plt flag.

relro_level: RelroLevel

Either partial, full, or off. Full RELRO makes the dynamic linker resolve all symbols at startup and marks the GOT read-only before starting the program, preventing overwriting the GOT.

archive_format: String

Format that archives should be emitted in. This affects whether we use LLVM to assemble an archive or fall back to the system linker, and currently only "gnu" is used to fall into LLVM. Unknown strings cause the system linker to be used.

allow_asm: bool

Is asm!() allowed? Defaults to true.

custom_unwind_resume: bool

Whether the target uses a custom unwind resumption routine. By default LLVM lowers resume instructions into calls to _Unwind_Resume defined in libgcc. If this option is enabled, the target must provide eh_unwind_resume lang item.

has_elf_tls: bool

Flag indicating whether ELF TLS (e.g., #[thread_local]) is available for this target.

obj_is_bitcode: boolno_integrated_as: boolmin_atomic_width: Option<u64>

Don't use this field; instead use the .min_atomic_width() method.

max_atomic_width: Option<u64>

Don't use this field; instead use the .max_atomic_width() method.

atomic_cas: bool

Whether the target supports atomic CAS operations natively

panic_strategy: PanicStrategy

Panic strategy: "unwind" or "abort"

abi_blacklist: Vec<Abi>

A blacklist of ABIs unsupported by the current target. Note that generic ABIs are considered to be supported on all platforms and cannot be blacklisted.

crt_static_allows_dylibs: bool

Whether or not linking dylibs to a static CRT is allowed.

crt_static_default: bool

Whether or not the CRT is statically linked by default.

crt_static_respected: bool

Whether or not crt-static is respected by the compiler (or is a no-op).

stack_probes: bool

Whether or not stack probes (__rust_probestack) are enabled

min_global_align: Option<u64>

The minimum alignment for global symbols.

default_codegen_units: Option<u64>

Default number of codegen units to use in debug mode

trap_unreachable: bool

Whether to generate trap instructions in places where optimization would otherwise produce control flow that falls through into unrelated memory.

requires_lto: bool

This target requires everything to be compiled with LTO to emit a final executable, aka there is no native linker for this target.

singlethread: bool

This target has no support for threads.

no_builtins: bool

Whether library functions call lowering/optimization is disabled in LLVM for this target unconditionally.

i128_lowering: bool

Whether to lower 128-bit operations to compiler_builtins calls. Use if your backend only supports 64-bit and smaller math.

codegen_backend: String

The codegen backend to use for this target, typically "llvm"

default_hidden_visibility: bool

The default visibility for symbols in this target should be "hidden" rather than "default"

embed_bitcode: bool

Whether or not bitcode is embedded in object files

emit_debug_gdb_scripts: bool

Whether a .debug_gdb_scripts section will be added to the output object file

requires_uwtable: bool

Whether or not to unconditionally uwtable attributes on functions, typically because the platform needs to unwind for things like stack unwinders.

simd_types_indirect: bool

Whether or not SIMD types are passed by reference in the Rust ABI, typically required if a target can be compiled with a mixed set of target features. This is true by default, and false for targets like wasm32 where the whole program either has simd or not.

override_export_symbols: Option<Vec<String>>

If set, have the linker export exactly these symbols, instead of using the usual logic to figure this out from the crate itself.

merge_functions: MergeFunctions

Determines how or whether the MergeFunctions LLVM pass should run for this target. Either "disabled", "trampolines", or "aliases". The MergeFunctions pass is generally useful, but some targets may need to opt out. The default is "aliases".

Workaround for: https://github.com/rust-lang/rust/issues/57356