Module rustc_target::spec::targets::wasm64_unknown_unknown

A “bare wasm” target representing a WebAssembly output that makes zero assumptions about its environment.

The wasm64-unknown-unknown target is intended to encapsulate use cases that do not rely on any imported functionality. The binaries generated are entirely self-contained by default when using the standard library. Although the standard library is available, most of it returns an error immediately (e.g. trying to create a TCP stream or something like that).

Functions

target