rustc_target

Module target_features

Declares Rust’s target feature names for each target. Note that these are similar to but not always identical to LLVM’s feature names, and Rust adds some features that do not correspond to LLVM features at all.

Structs

FeatureConstraints

Enums

Stability

Stability information for target features.

Constants

AARCH64_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

AARCH64_TIED_FEATURES 🔒

ARM_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

BPF_FEATURES 🔒

CSKY_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

HEXAGON_FEATURES 🔒

HEXAGON_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

IBMZ_FEATURES 🔒

LOONGARCH_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

MIPS_FEATURES 🔒

MIPS_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

NVPTX_FEATURES 🔒

POWERPC_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

RISCV_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

RUSTC_SPECIFIC_FEATURES

Features that control behaviour of rustc, rather than the codegen. These exist globally and are not in the target-specific lists below.

S390X_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

SPARC_FEATURES 🔒

SPARC_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

WASM_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

X86_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

Statics

AARCH64_FEATURES 🔒

ARM_FEATURES 🔒

CSKY_FEATURES 🔒

LOONGARCH_FEATURES 🔒

M68K_FEATURES 🔒

POWERPC_FEATURES 🔒

RISCV_FEATURES 🔒

WASM_FEATURES 🔒

X86_FEATURES 🔒

Functions

all_rust_features

When rustdoc is running, provide a list of all known features so that all their respective primitives may be documented.

Type Aliases

ImpliedFeatures 🔒