Module rustc_target::abi

Re-exports

pub use Integer::*;

pub use Primitive::*;

Modules

call

Structs

AbiAndPrefAlign
A pair of aligments, ABI-mandated and preferred.

Align
Alignment of a type in bytes (always a power of two).

LayoutDetails

Scalar
Information about one scalar component of a Rust type.

Size
Size of a type in bytes.

TargetDataLayout
Parsed Data layout for a target, which contains everything needed to compute layouts.

TyLayout
The details of the layout of a type, alongside the type itself. Provides various type traversal APIs (e.g., recursing into fields).

VariantIdx

Enums

Abi
Describes how values of the type are passed by target ABIs, in terms of categories of C types there are ABI rules for.

Endian
Endianness of the target, which must match cfg(target-endian).

FieldPlacement
Describes how the fields of a type are located in memory.

FloatTy

Integer
Integers, also used for enum discriminants.

Primitive
Fundamental unit of memory access and layout.

Variants

Traits

HasDataLayout

LayoutOf

TyLayoutMethods