Struct rustc_target::abi::TargetDataLayout

pub struct TargetDataLayout {
    pub endian: Endian,
    pub i1_align: AbiAndPrefAlign,
    pub i8_align: AbiAndPrefAlign,
    pub i16_align: AbiAndPrefAlign,
    pub i32_align: AbiAndPrefAlign,
    pub i64_align: AbiAndPrefAlign,
    pub i128_align: AbiAndPrefAlign,
    pub f32_align: AbiAndPrefAlign,
    pub f64_align: AbiAndPrefAlign,
    pub pointer_size: Size,
    pub pointer_align: AbiAndPrefAlign,
    pub aggregate_align: AbiAndPrefAlign,
    pub vector_align: Vec<(Size, AbiAndPrefAlign)>,
    pub instruction_address_space: u32,
}

Parsed Data layout for a target, which contains everything needed to compute layouts.

Fields
endian: Endiani1_align: AbiAndPrefAligni8_align: AbiAndPrefAligni16_align: AbiAndPrefAligni32_align: AbiAndPrefAligni64_align: AbiAndPrefAligni128_align: AbiAndPrefAlignf32_align: AbiAndPrefAlignf64_align: AbiAndPrefAlignpointer_size: Sizepointer_align: AbiAndPrefAlignaggregate_align: AbiAndPrefAlignvector_align: Vec<(Size, AbiAndPrefAlign)>

Alignments for vector types.

instruction_address_space: u32
Methods

impl TargetDataLayout

pub fn parse(target: &Target) -> Result<TargetDataLayout, String>

pub fn obj_size_bound(&self) -> u64

Returns exclusive upper bound on object size.

The theoretical maximum object size is defined as the maximum positive isize value. This ensures that the offset semantics remain well-defined by allowing it to correctly index every address within an object along with one byte past the end, along with allowing isize to store the difference between any two pointers into an object.

The upper bound on 64-bit currently needs to be lower because LLVM uses a 64-bit integer to represent object size in bits. It would need to be 1 << 61 to account for this, but is currently conservatively bounded to 1 << 47 as that is enough to cover the current usable address space on 64-bit ARMv8 and x86_64.

pub fn ptr_sized_integer(&self) -> Integer

pub fn vector_align(&self, vec_size: Size) -> AbiAndPrefAlign