rustc_target
Module asm

Modules

aarch64 🔒

arm 🔒

avr 🔒

bpf 🔒

csky 🔒

hexagon 🔒

loongarch 🔒

m68k 🔒

mips 🔒

msp430 🔒

nvptx 🔒

powerpc 🔒

riscv 🔒

s390x 🔒

spirv 🔒

wasm 🔒

x86 🔒

Macros

def_reg_class 🔒

def_regs 🔒

types 🔒

Structs

ModifierInfo

Enums

AArch64InlineAsmReg

AArch64InlineAsmRegClass

ArmInlineAsmReg

ArmInlineAsmRegClass

AvrInlineAsmReg

AvrInlineAsmRegClass

BpfInlineAsmReg

BpfInlineAsmRegClass

CSKYInlineAsmReg

CSKYInlineAsmRegClass

HexagonInlineAsmReg

HexagonInlineAsmRegClass

InlineAsmArch

InlineAsmClobberAbi

InlineAsmReg

InlineAsmRegClass

InlineAsmRegOrRegClass

InlineAsmType

Set of types which can be used with a particular register class.

LoongArchInlineAsmReg

LoongArchInlineAsmRegClass

M68kInlineAsmReg

M68kInlineAsmRegClass

MipsInlineAsmReg

MipsInlineAsmRegClass

Msp430InlineAsmReg

Msp430InlineAsmRegClass

NvptxInlineAsmReg

NvptxInlineAsmRegClass

PowerPCInlineAsmReg

PowerPCInlineAsmRegClass

RiscVInlineAsmReg

RiscVInlineAsmRegClass

S390xInlineAsmReg

S390xInlineAsmRegClass

SpirVInlineAsmReg

SpirVInlineAsmRegClass

WasmInlineAsmReg

WasmInlineAsmRegClass

X86InlineAsmReg

X86InlineAsmRegClass

Functions

allocatable_registers

Returns the full set of allocatable registers for a given architecture.