Module rustc_target::asm

Modules

aarch64 🔒

arm 🔒

avr 🔒

bpf 🔒

hexagon 🔒

mips 🔒

msp430 🔒

nvptx 🔒

powerpc 🔒

riscv 🔒

s390x 🔒

spirv 🔒

wasm 🔒

x86 🔒

Macros

def_reg_class 🔒

def_regs 🔒

types 🔒

Enums

AArch64InlineAsmReg

AArch64InlineAsmRegClass

ArmInlineAsmReg

ArmInlineAsmRegClass

AvrInlineAsmReg

AvrInlineAsmRegClass

BpfInlineAsmReg

BpfInlineAsmRegClass

HexagonInlineAsmReg

HexagonInlineAsmRegClass

InlineAsmArch

InlineAsmClobberAbi

InlineAsmReg

InlineAsmRegClass

InlineAsmRegOrRegClass

InlineAsmType

Set of types which can be used with a particular register class.

MipsInlineAsmReg

MipsInlineAsmRegClass

Msp430InlineAsmReg

Msp430InlineAsmRegClass

NvptxInlineAsmReg

NvptxInlineAsmRegClass

PowerPCInlineAsmReg

PowerPCInlineAsmRegClass

RiscVInlineAsmReg

RiscVInlineAsmRegClass

S390xInlineAsmReg

S390xInlineAsmRegClass

SpirVInlineAsmReg

SpirVInlineAsmRegClass

WasmInlineAsmReg

WasmInlineAsmRegClass

X86InlineAsmReg

X86InlineAsmRegClass

Functions

allocatable_registers

Returns the full set of allocatable registers for a given architecture.