Struct rustc_target::spec::Target

pub struct Target {
    pub llvm_target: Cow<'static, str>,
    pub metadata: TargetMetadata,
    pub pointer_width: u32,
    pub arch: Cow<'static, str>,
    pub data_layout: Cow<'static, str>,
    pub options: TargetOptions,
}

Everything rustc knows about how to compile for a specific target.

Every field here must be specified, and has no default value.

Fields
llvm_target: Cow<'static, str>

Target triple to pass to LLVM.

metadata: TargetMetadata

Metadata about a target, for example the description or tier. Used for generating target documentation.

pointer_width: u32

Number of bits in a pointer. Influences the target_pointer_width cfg variable.

arch: Cow<'static, str>

Architecture to use for ABI considerations. Valid options include: โx86โ, โx86_64โ, โarmโ, โaarch64โ, โmipsโ, โpowerpcโ, โpowerpc64โ, and others.

data_layout: Cow<'static, str>

Data layout to pass to LLVM.

options: TargetOptions

Optional settings with defaults.

Implementations

impl Target

pub fn parse_data_layout( &self, ) -> Result<TargetDataLayout, TargetDataLayoutErrors<'_>>

impl Target

pub fn adjust_abi(&self, abi: Abi, c_variadic: bool) -> Abi

Given a function ABI, turn it into the correct ABI for this target.

pub fn is_abi_supported(&self, abi: Abi) -> Option<bool>

Returns a None if the UNSUPPORTED_CALLING_CONVENTIONS lint should be emitted

pub fn min_atomic_width(&self) -> u64

Minimum integer size in bits that this target can perform atomic operations on.

pub fn max_atomic_width(&self) -> u64

Maximum integer size in bits that this target can perform atomic operations on.

pub fn from_json(obj: Json) -> Result<(Target, TargetWarnings), String>

Loads a target descriptor from a JSON object.

pub fn expect_builtin(target_triple: &TargetTriple) -> Target

Load a built-in target

pub fn search( target_triple: &TargetTriple, sysroot: &Path, ) -> Result<(Target, TargetWarnings), String>

Search for a JSON file specifying the given target triple.

If none is found in $RUST_TARGET_PATH, look for a file called target.json inside the sysroot under the target-tripleโs rustlib directory. Note that it could also just be a bare filename already, so also check for that. If one of the hardcoded targets we know about, just return it directly.

The error string could come from any of the APIs called, including filesystem access and JSON decoding.

impl Target

pub fn supported_target_features(&self) -> &'static [(&'static str, Stability)]

pub fn tied_target_features(&self) -> &'static [&'static [&'static str]]