Struct rustc_target::spec::Target

pub struct Target {
    pub llvm_target: String,
    pub target_endian: String,
    pub target_pointer_width: String,
    pub target_c_int_width: String,
    pub target_os: String,
    pub target_env: String,
    pub target_vendor: String,
    pub arch: String,
    pub data_layout: String,
    pub linker_flavor: LinkerFlavor,
    pub options: TargetOptions,
}

Everything rustc knows about how to compile for a specific target.

Every field here must be specified, and has no default value.

Fields
llvm_target: String

Target triple to pass to LLVM.

target_endian: String

String to use as the target_endian cfg variable.

target_pointer_width: String

String to use as the target_pointer_width cfg variable.

target_c_int_width: String

Width of c_int type

target_os: String

OS name to use for conditional compilation.

target_env: String

Environment name to use for conditional compilation.

target_vendor: String

Vendor name to use for conditional compilation.

arch: String

Architecture to use for ABI considerations. Valid options: "x86", "x86_64", "arm", "aarch64", "mips", "powerpc", and "powerpc64".

data_layout: String

Data layout to pass to LLVM.

linker_flavor: LinkerFlavor

Linker flavor

options: TargetOptions

Optional settings with defaults.

Methods

impl Target

pub fn adjust_abi(&self, abi: Abi) -> Abi

Given a function ABI, turn it into the correct ABI for this target.

pub fn min_atomic_width(&self) -> u64

Minimum integer size in bits that this target can perform atomic operations on.

pub fn max_atomic_width(&self) -> u64

Maximum integer size in bits that this target can perform atomic operations on.

pub fn is_abi_supported(&self, abi: Abi) -> bool

pub fn from_json(obj: Json) -> TargetResult

Loads a target descriptor from a JSON object.

pub fn search(target_triple: &TargetTriple) -> Result<Target, String>

Search RUST_TARGET_PATH for a JSON file specifying the given target triple. Note that it could also just be a bare filename already, so also check for that. If one of the hardcoded targets we know about, just return it directly.

The error string could come from any of the APIs called, including filesystem access and JSON decoding.