rustc_target
Module target_features

Declares Rust’s target feature names for each target. Note that these are similar to but not always identical to LLVM’s feature names, and Rust adds some features that do not correspond to LLVM features at all.

Enums

Stability

Stability information for target features. Toggleability is the type storing whether (un)stable features can be toggled: this is initially a function since it can depend on Target, but for stable hashing it needs to be something hashable to we have to make the type generic.

Constants

AARCH64_FEATURES 🔒

AARCH64_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

AARCH64_TIED_FEATURES 🔒

ARM_FEATURES 🔒

ARM_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

BPF_FEATURES 🔒

CSKY_FEATURES 🔒

CSKY_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

HEXAGON_FEATURES 🔒

HEXAGON_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

IBMZ_FEATURES 🔒

LOONGARCH_FEATURES 🔒

LOONGARCH_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

MIPS_FEATURES 🔒

MIPS_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

POWERPC_FEATURES 🔒

POWERPC_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

RISCV_FEATURES 🔒

RISCV_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

RUSTC_SPECIAL_FEATURES

Features that require special handling when passing to LLVM: these are target-specific (i.e., must also be listed in the target-specific list below) but do not correspond to an LLVM target feature.

RUSTC_SPECIFIC_FEATURES

Features that control behaviour of rustc, rather than the codegen. These exist globally and are not in the target-specific lists below.

S390X_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

SPARC_FEATURES 🔒

SPARC_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

STABLE 🔒

WASM_FEATURES 🔒

WASM_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

X86_FEATURES 🔒

X86_FEATURES_FOR_CORRECT_VECTOR_ABI 🔒

Functions

all_rust_features

When rustdoc is running, provide a list of all known features so that all their respective primitives may be documented.

unstable 🔒

Type Aliases

ImpliedFeatures 🔒

StabilityComputed

Stability where allow_toggle has already been computed.

StabilityUncomputed

Stability where allow_toggle has not been computed yet. Returns Ok if the toggle is allowed, Err with an explanation of not.