rustc_target
Module target_features

Enums

Stability

Stability information for target features.

Constants

AARCH64_ALLOWED_FEATURES 🔒

AARCH64_TIED_FEATURES 🔒

ARM_ALLOWED_FEATURES 🔒

BPF_ALLOWED_FEATURES 🔒

CSKY_ALLOWED_FEATURES 🔒

HEXAGON_ALLOWED_FEATURES 🔒

IBMZ_ALLOWED_FEATURES 🔒

LOONGARCH_ALLOWED_FEATURES 🔒

MIPS_ALLOWED_FEATURES 🔒

POWERPC_ALLOWED_FEATURES 🔒

RISCV_ALLOWED_FEATURES 🔒

RUSTC_SPECIAL_FEATURES

Features that require special handling when passing to LLVM.

RUSTC_SPECIFIC_FEATURES

Features that control behaviour of rustc, rather than the codegen.

WASM_ALLOWED_FEATURES 🔒

X86_ALLOWED_FEATURES 🔒

Functions

all_known_features

When rustdoc is running, provide a list of all known features so that all their respective primitives may be documented.

Type Aliases

ImpliedFeatures 🔒